# apps/comment-service/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/v1/health": {
      "get": {
        "operationId": "HealthController_health",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Health"
        ]
      }
    },
    "/v1/local/auth/operator/login": {
      "post": {
        "operationId": "LocalAuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocalLoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "LocalAuth"
        ]
      }
    },
    "/v1/local/auth/member/token": {
      "post": {
        "operationId": "LocalAuthController_memberToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueMemberTokenDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "LocalAuth"
        ]
      }
    },
    "/v1/console/applications": {
      "post": {
        "operationId": "ApplicationsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApplicationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Applications"
        ]
      },
      "get": {
        "operationId": "ApplicationsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Applications"
        ]
      }
    },
    "/v1/console/applications/{key}": {
      "patch": {
        "operationId": "ApplicationsController_update",
        "parameters": [
          {
            "name": "key",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateApplicationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Applications"
        ]
      }
    },
    "/v1/articles/{articleKey}/comments": {
      "post": {
        "operationId": "CommentsController_create",
        "parameters": [
          {
            "name": "x-application-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "articleKey",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Comments"
        ]
      },
      "get": {
        "operationId": "CommentsController_list",
        "parameters": [
          {
            "name": "x-application-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "articleKey",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Comments"
        ]
      }
    },
    "/v1/comments/{commentId}/replies": {
      "post": {
        "operationId": "CommentsController_reply",
        "parameters": [
          {
            "name": "x-application-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Comments"
        ]
      }
    },
    "/v1/comments/{commentId}/branch": {
      "get": {
        "operationId": "CommentsController_branch",
        "parameters": [
          {
            "name": "x-application-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Comments"
        ]
      }
    },
    "/v1/comments/{commentId}/reactions/{emoji}": {
      "put": {
        "operationId": "ReactionsController_toggle",
        "parameters": [
          {
            "name": "x-application-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Reactions"
        ]
      }
    },
    "/v1/comments/{commentId}/triple-reaction": {
      "post": {
        "operationId": "ReactionsController_triple",
        "parameters": [
          {
            "name": "x-application-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Reactions"
        ]
      }
    },
    "/v1/console/settings": {
      "get": {
        "operationId": "SettingsController_get",
        "parameters": [
          {
            "name": "x-application-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Settings"
        ]
      },
      "put": {
        "operationId": "SettingsController_update",
        "parameters": [
          {
            "name": "x-application-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSettingsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Settings"
        ]
      }
    },
    "/v1/console/comments": {
      "get": {
        "operationId": "ConsoleCommentsController_list",
        "parameters": [
          {
            "name": "x-application-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ConsoleComments"
        ]
      }
    },
    "/v1/console/comments/{commentId}": {
      "delete": {
        "operationId": "ConsoleCommentsController_delete",
        "parameters": [
          {
            "name": "x-application-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "ConsoleComments"
        ]
      }
    },
    "/v1/users/{memberId}/mute": {
      "put": {
        "operationId": "MutesController_mute",
        "parameters": [
          {
            "name": "x-application-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "memberId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Mutes"
        ]
      },
      "delete": {
        "operationId": "MutesController_unmute",
        "parameters": [
          {
            "name": "x-application-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "memberId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Mutes"
        ]
      }
    },
    "/v1/comments/{commentId}/reports": {
      "post": {
        "operationId": "ReportsController_create",
        "parameters": [
          {
            "name": "x-application-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReportDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Reports"
        ]
      }
    },
    "/v1/console/users/{memberId}/block": {
      "put": {
        "operationId": "BlocksController_set",
        "parameters": [
          {
            "name": "x-application-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "memberId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetBlockDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Blocks"
        ]
      },
      "delete": {
        "operationId": "BlocksController_remove",
        "parameters": [
          {
            "name": "x-application-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "memberId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Blocks"
        ]
      }
    }
  },
  "info": {
    "title": "Comment API",
    "description": "",
    "version": "v1",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "LocalLoginDto": {
        "type": "object",
        "properties": {}
      },
      "IssueMemberTokenDto": {
        "type": "object",
        "properties": {}
      },
      "CreateApplicationDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateApplicationDto": {
        "type": "object",
        "properties": {}
      },
      "CreateCommentDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateSettingsDto": {
        "type": "object",
        "properties": {}
      },
      "CreateReportDto": {
        "type": "object",
        "properties": {}
      },
      "SetBlockDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}